+++ test/Filter/ODMFilterTest.php
@@ -48,16 +48,19 @@ class ODMFilterTest extends TestCase
 
         $meta1 = new Document\Meta;
         $meta1->setName('MetaOne');
+        $meta1->setDescription('Foo');
         $meta1->setCreatedAt(new DateTime('2011-12-18 13:17:17'));
         $objectManager->persist($meta1);
 
         $meta2 = new Document\Meta;
         $meta2->setName('MetaTwo');
+        $meta2->setDescription('Bar');
         $meta2->setCreatedAt(new DateTime('2014-12-18 13:17:17'));
         $objectManager->persist($meta2);
 
         $meta3 = new Document\Meta;
         $meta3->setName('MetaThree');
+        $meta3->setDescription('Baz');
         $meta3->setCreatedAt(new DateTime('2012-12-18 13:17:17'));
         $objectManager->persist($meta3);
 
@@ -370,7 +373,6 @@ class ODMFilterTest extends TestCase
 
         $this->assertEquals(1, $this->countResult($filters));
 
-
         $filters = [
             [
                 'field' => 'createdAt',
@@ -381,6 +383,29 @@ class ODMFilterTest extends TestCase
 
         $this->assertEquals(1, $this->countResult($filters));
 
+        $filters = [
+            [
+                'field' => 'description',
+                'type' => 'isnull',
+            ],
+        ];
+
+        $this->assertEquals(2, $this->countResult($filters));
+
+        $filters = [
+            [
+                'field' => 'description',
+                'where' => 'and',
+                'type' => 'isnull',
+            ],
+            [
+                'field' => 'createdAt',
+                'where' => 'and',
+                'type' => 'isnull',
+            ],
+        ];
+
+        $this->assertEquals(1, $this->countResult($filters));
 
         $filters = [
             [
@@ -397,6 +422,15 @@ class ODMFilterTest extends TestCase
         ];
 
         $this->assertEquals(2, $this->countResult($filters));
+
+        $filters = [
+            [
+                'field' => 'nonExistingField',
+                'type' => 'isnull',
+            ],
+        ];
+
+        $this->assertEquals(5, $this->countResult($filters));
     }
 
     public function testIsNotNull()
@@ -410,7 +444,6 @@ class ODMFilterTest extends TestCase
 
         $this->assertEquals(4, $this->countResult($filters));
 
-
         $filters = [
             [
                 'field' => 'createdAt',
@@ -421,6 +454,29 @@ class ODMFilterTest extends TestCase
 
         $this->assertEquals(4, $this->countResult($filters));
 
+        $filters = [
+            [
+                'field' => 'description',
+                'type' => 'isnotnull',
+            ],
+        ];
+
+        $this->assertEquals(3, $this->countResult($filters));
+
+        $filters = [
+            [
+                'field' => 'description',
+                'where' => 'and',
+                'type' => 'isnotnull',
+            ],
+            [
+                'field' => 'createdAt',
+                'where' => 'and',
+                'type' => 'isnotnull',
+            ],
+        ];
+
+        $this->assertEquals(3, $this->countResult($filters));
 
         $filters = [
             [
@@ -437,6 +493,15 @@ class ODMFilterTest extends TestCase
         ];
 
         $this->assertEquals(5, $this->countResult($filters));
+
+        $filters = [
+            [
+                'field' => 'nonExistingField',
+                'type' => 'isnotnull',
+            ],
+        ];
+
+        $this->assertEquals(0, $this->countResult($filters));
     }
 
     public function testIn()